mod dao daodb daoimpl

+++ src/main/java/com/practice/webapp/dao/MemberDAO.java
@@ -6,9 +6,9 @@ import com.practice.webapp.entity.Member;
 public interface MemberDAO {
 	public List<Member> getList();
 	public void insert(Member member);
-	public Member get(long id);
+	public Member get(int id);
     public void update(Member member,int M_id);
-    public void delete(long id);
+    public void delete(int id);
 	public boolean checkLoginMember(String M_idName, String password);
 //	public int count();
